feat: Create TCL script for sending files to replace existing files
The script is intended to overcome a limitation on the remote server that prevents deleting existing files. Remove files can be renamed, then later deleted (or deleted via cron), but they cannot be directly replaced via SFTP. The script overcomes this limitation by renaming the old copy on the remote server to put it in a `.trash` directory relative to the user's home directory. It uses a timestamp to avoid name conflicts.